feat: 설명 수만큼 Sector의 Description 을 가져오는 기능 추가
